+++ packages/common/src/models/Draw/MatchType.ts
@@ -1,0 +1,4 @@
+import { RegularMatchType } from '../Regular';
+import { LotteryMatchType } from '../Lottery';
+
+export type MatchType = RegularMatchType | LotteryMatchType;