+++ packages/common/src/models/Lottery/LotteryMatchType.ts
@@ -1,0 +1,23 @@
+export enum LotteryMatchType {
+  $7_RAKAM = '$7_RAKAM',
+  $6_RAKAM = '$6_RAKAM',
+  SON_ALTI_RAKAM = 'SON_ALTI_RAKAM',
+  SON_BES_RAKAM = 'SON_BES_RAKAM',
+  SON_DORT_RAKAM = 'SON_DORT_RAKAM',
+  SON_UC_RAKAM = 'SON_UC_RAKAM',
+  SON_IKI_RAKAM = 'SON_IKI_RAKAM',
+  AMORTI = 'AMORTI',
+  TESELLI = 'TESELLI',
+}
+
+export const LotteryMatchTypeMap: { [key: string]: string } = {
+  $7_RAKAM: '7 Rakam',
+  $6_RAKAM: '6 Rakam',
+  SON_ALTI_RAKAM: 'Son 6 Rakam',
+  SON_BES_RAKAM: 'Son 5 Rakam',
+  SON_DORT_RAKAM: 'Son 4 Rakam',
+  SON_UC_RAKAM: 'Son 3 Rakam',
+  SON_IKI_RAKAM: 'Son 2 Rakam',
+  AMORTI: 'Amorti',
+  TESELLI: 'Teselli',
+};