+++ packages/common/src/models/Lottery/LotteryReportLine.ts
@@ -1,0 +1,9 @@
+import { IReportLine } from '../ProcessDraw';
+
+/**
+ * Interface definiton of report lines for lottery games.
+ * Notice that the winningNumbers field is of type string[]!
+ */
+export interface LotteryReportLine extends IReportLine {
+  winningNumbers: string[];
+}