+++ packages/common/src/models/Lottery/index.ts
@@ -1,4 +1,5 @@
 export * from './LotteryCategory';
 export * from './LotteryDrawData';
 export * from './LotteryGame';
-export * from './MatchTypeLottery';
+export * from './LotteryMatchType';
+export * from './LotteryReportLine';