+++ packages/common/src/models/ProcessDraw/IReportLine.ts
@@ -1,0 +1,7 @@
+/**
+ * Base definiton for report lines.
+ */
+export interface IReportLine {
+  type: string;
+  prize: number;
+}