+++ packages/common/src/models/ProcessDraw/ReportLineType.ts
@@ -1,0 +1,4 @@
+import { RegularReportLine } from '../Regular';
+import { LotteryReportLine } from '../Lottery';
+
+export type ReportLineType = RegularReportLine | LotteryReportLine;