+++ packages/common/src/models/ProcessDraw/index.ts
@@ -1,1 +1,3 @@
+export * from './IReportLine';
 export * from './ProcessDraw';
+export * from './ReportLineType';