+++ packages/common/src/models/Regular/RegularReportLine.ts
@@ -1,0 +1,8 @@
+import { IReportLine } from '../ProcessDraw';
+
+/**
+ * Interface definiton of report lines for regular games.
+ */
+export interface RegularReportLine extends IReportLine {
+  winnersCount: number;
+}