+++ packages/common/src/models/Regular/index.ts
@@ -1,5 +1,6 @@
-export * from './MatchTypeRegular';
 export * from './RegularDrawData';
 export * from './RegularDrawType';
 export * from './RegularGame';
+export * from './RegularMatchType';
+export * from './RegularReportLine';
 export * from './WinnerCategory';