+++ packages/common/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.3.0] - 2020-04-02
+
+### Added
+
+- ProcessDraw.report() implemented
+- DrawUtils.matchTypeToString() implemented
+
 ## [2.2.0] - 2020-04-02
 
 ### Removed